fix: move recommended artists functions
